Upgrade yaml to v3
In yaml.v3 there is a change for string-keyed yaml objects in
that they actually become map[string]any instead of map[any]any where
the yaml keys in that object are all strings (for example a yaml object
with an integer as a key will still have map[any]any).

Because the open api spec is json it means that the spec can never have
non-string keys, this cleans up the code ever so slightly because we can
now use map[string]any instead of map[any]any which is nice.